+++ src/main/kotlin/icu/windea/pls/integrations/lint/PlsLintManager.kt
@@ -1,4 +1,11 @@
 package icu.windea.pls.integrations.lint
 
+import icu.windea.pls.core.collections.optimized
+import icu.windea.pls.integrations.lint.tools.*
+import icu.windea.pls.model.*
+
 object PlsLintManager {
+    fun findTools(gameType: ParadoxGameType?): List<PlsLintToolProvider> {
+        return PlsLintToolProvider.EP_NAME.extensions.filter { it.isEnabled() && it.isAvailable(gameType) && it.isValid() }.optimized()
+    }
 }