+++ src/main/kotlin/icu/windea/pls/integrations/lint/PlsLintResult.kt
@@ -1,0 +1,3 @@
+package icu.windea.pls.integrations.lint
+
+interface PlsLintResult