+++ src/main/kotlin/icu/windea/pls/integrations/lint/tools/PlsCommandBasedLintToolProvider.kt
@@ -1,0 +1,4 @@
+package icu.windea.pls.integrations.lint.tools
+
+abstract class PlsCommandBasedLintToolProvider: PlsLintToolProvider
+