+++ src/main/kotlin/icu/windea/pls/integrations/lint/tools/PlsLintToolProvider.kt
@@ -1,0 +1,29 @@
+package icu.windea.pls.integrations.lint.tools
+
+import com.intellij.openapi.extensions.*
+import com.intellij.openapi.vfs.*
+import icu.windea.pls.core.annotations.*
+import icu.windea.pls.integrations.lint.*
+import icu.windea.pls.model.ParadoxGameType
+
+/**
+ * 提供检查工具。用于提供额外的代码检查。
+ *
+ * 注意：具体的操作方法可能不会再次验证工具是否可用且合法。
+ */
+interface PlsLintToolProvider {
+    fun isEnabled(): Boolean
+
+    fun isAvailable(gameType: ParadoxGameType?): Boolean
+
+    fun isValid(): Boolean
+
+    fun validateFile(file: VirtualFile): PlsLintResult?
+
+    fun validateRootDirectory(rootDirectory: VirtualFile): PlsLintResult?
+
+    companion object INSTANCE {
+        val EP_NAME = ExtensionPointName<PlsLintToolProvider>("icu.windea.pls.integrations.lintToolProvider")
+    }
+}
+