+++ src/main/kotlin/icu/windea/pls/integrations/translation/tools/PlsTranslationToolProvider.kt
@@ -5,7 +5,7 @@ import icu.windea.pls.config.config.*
 import icu.windea.pls.integrations.translation.*
 
 /**
- * 提供翻译工具，用于翻译本地化文本。
+ * 提供翻译工具。用于翻译本地化文本。
  *
  * 注意：具体的操作方法不会再次验证工具是否受支持（[supports]）。
  */